fix: 구동 검증 및 Q-Sign 인프라 레이어 완성 v1.4.1

+++ q-sign/src/main/java/kr/go/smes/qsign/outbox/QSignOutboxRecord.java
@@ -46,7 +46,7 @@ public class QSignOutboxRecord {
     private String status;   // PENDING / PUBLISHED / FAILED
 
     @Column(name = "retry_count", nullable = false)
-    private int retryCount;
+    private short retryCount;  // DB: SMALLINT (V1 migration 기준)
 
     @Column(name = "error_message", columnDefinition = "text")
     private String errorMessage;
